fix type conversion error (on windows) unsigned/signed affecting timestamps on rrd_dump.
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.2/program@1744 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_dump.c
@@ -318,7 +318,7 @@ rrd_dump_r(const char *filename, char *outname)
 		       +(rrd.rra_ptr[i].cur_row+1)
 		       * rrd.stat_head->ds_cnt
 		       * sizeof(rrd_value_t)),SEEK_SET);
-	timer = - (rrd.rra_def[i].row_cnt-1);
+	timer = - (long)(rrd.rra_def[i].row_cnt-1);
 	ii=rrd.rra_ptr[i].cur_row;
 	for(ix=0;ix<rrd.rra_def[i].row_cnt;ix++){	    
 	    ii++;
